chore: fix tmq meta bug
