+++ deploy-pom.xml
@@ -5,7 +5,7 @@
 
     <groupId>com.taosdata.jdbc</groupId>
     <artifactId>taos-jdbcdriver</artifactId>
-    <version>3.8.2</version>
+    <version>3.8.3</version>
 
     <packaging>jar</packaging>
